pyproject.toml
[project]
name = "uibar-project-new"
version = "0.1.0"
description = "A Django project for accommodation booking."
authors = [
    {name = "Gulnara",email = "gulnarkaz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt[blacklist] (>=5.5.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
